+++ colorSetting.js
@@ -496,7 +496,7 @@ async function getEyeDropper(event, _this) {
 
         // rerender canvas for fresh data
         canvas.app.render();
-        let imageCanvas = canvas.app.renderer.plugins.extract.canvas(canvas.stage);
+        let imageCanvas = canvas.app.renderer.extract?.canvas(canvas.stage) ?? canvas.app.renderer.plugins.extract.canvas(canvas.stage);
         let ctx = imageCanvas.getContext('2d');
         let imageData = ctx.getImageData(x, y, 1, 1).data;
         let color = [imageData[0], imageData[1], imageData[2], imageData[3] / 255];
@@ -638,53 +638,6 @@ Hooks.once('init', function () {
     registerInitVar();
     registerButton();
     Hooks.callAll("colorSettingsInitialized", window.Ardittristan.ColorSetting);
-});
-
-
-Hooks.once('init', function () {
-    game.settings.register("colorsettings", "showWarning", {
-        config: true,
-        type: Boolean,
-        default: true,
-        name: "Show Error",
-        hint: "Enable or disable error if main module missing."
-    });
-    if (window?.Ardittristan?.initialColorSettingRun === undefined) {
-        if (game.modules.has("colorsettings")) {
-            if (game.modules.get("colorsettings").active) {
-                return;
-            } else {
-                game.settings.register("colorsettings", "autoEnable", {
-                    config: false,
-                    type: Boolean,
-                    default: true
-                });
-                Hooks.once("canvasReady", () => {
-                    if (game.user.isGM && game.settings.get("colorsettings", "autoEnable")) {
-                        Dialog.confirm({
-                            title: "Enable Color Settings module?",
-                            content: "<p>You seem to have Color Settings installed already, do you want to enable it?</p>",
-                            yes: () => game.settings.set("core", ModuleManagement.CONFIG_SETTING, {
-                                ...game.settings.get("core", ModuleManagement.CONFIG_SETTING),
-                                ...{ colorsettings: true }
-                            }),
-                            no: () => game.settings.set("colorsettings", "autoEnable", false),
-                            defaultYes: false
-                        });
-
-                    }
-                });
-            }
-        }
-
-        Hooks.once('ready', function () {
-            if (game.settings.get("colorsettings", "showWarning")) {
-                ui.notifications.notify("colorsettings | A module is missing the color picker library. For best results, please install and enable the Lib-Color Settings module.", "warning");
-            }
-        });
-        console.log("colorsettings | initializing fallback mode");
-
-    }
 });
 
 Hooks.once("setup", async function () {